requirements.txt
yt-dlp
openai
whisper
torch
colorama
python-dotenv

